Implement vault folder management by creating a dedicated vaultFolders storage module. Migrate related functions from the core storage interface, ensuring a cleaner architecture. Update references in both desktop and extension clients to utilize the new module.
